% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_lasso.R
\name{ADMM_lasso}
\alias{ADMM_lasso}
\title{implementation of ADMM algorithm with Lasso penalty}
\usage{
ADMM_lasso(A, b, lambda, rho, iter = 1000, e.abs = 1e-04, e.rel = 0.01)
}
\arguments{
\item{A}{design matrix}

\item{b}{response vector}

\item{lambda}{overrelaxation paramter}

\item{rho}{augmented Lagrangian parameter}

\item{iter}{max number of iterations}

\item{e.abs}{absolute tolerance stopping criterion}

\item{e.rel}{relative tolerance stopping criterion}
}
\value{
a list consisting of the iteration history
}
\description{
implementation of ADMM algorithm with Lasso penalty
}
\examples{
m <- 500 # number of samples
n <- 1000 # number of features
p <- 0.1 # sparsity density: percentage of non-zero elements

set.seed(0)
# generate a random sparse matrix
x0 <- matrix(Matrix::rsparsematrix(n,1,p))
A <- matrix(stats::rnorm(m*n),nrow=m)

# normalize columns
for (i in 1:ncol(A)){
  A[,i] <- A[,i]/sqrt(sum(A[,i]*A[,i]))
}

b <- A\%*\%x0 + sqrt(0.001)*matrix(stats::rnorm(m))

# set regularization term lambda
lambda <- 0.1*base::norm(t(A)\%*\%b, "F")

rho <- 1
ADMM_lasso(A,b,lambda,rho)
}
