% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkr_demo.R
\name{sparkr_demo}
\alias{sparkr_demo}
\title{a Spark example that fits a linear regression model to mtcars dataset
and disconnects the Spark cluster}
\usage{
sparkr_demo(sc)
}
\arguments{
\item{sc}{a Spark cluster (create by calling spark_connect)}
}
\value{
fitted linear regresion model
}
\description{
a Spark example that fits a linear regression model to mtcars dataset
and disconnects the Spark cluster
}
\examples{
sparkr_demo(sparklyr::spark_connect(master="local"))
}
